Allow running mount-tests on Windows

+++ test/mount_test/mount_test.go
@@ -45,6 +45,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"runtime"
 	"testing"
 	"time"
 
@@ -99,6 +100,13 @@ func blobfuseUnmount(suite *mountSuite, unmountOutput string) {
 
 // mount command test along with remount on the same path
 func (suite *mountSuite) TestMountCmd() {
+	// TODO: Allow test to run on Windows.
+	if runtime.GOOS == "windows" {
+		suite.Equal(0, 0)
+		fmt.Println("Skipping tests on windows. We should enable this in the future.")
+		return
+	}
+
 	// run mount command
 	mountCmd := exec.Command(blobfuseBinary, "mount", mntDir, "--config-file="+configFile)
 	cliOut, err := mountCmd.Output()
@@ -120,24 +128,45 @@ func (suite *mountSuite) TestMountCmd() {
 	blobfuseUnmount(suite, mntDir)
 }
 
-// mount failure test where the mount directory does not exists
+// mount failure test where the mount directory does not exists on Linux
+// or does exist on Windows
 func (suite *mountSuite) TestMountDirNotExists() {
-	tempDir := filepath.Join(mntDir, "tempdir")
-	mountCmd := exec.Command(blobfuseBinary, "mount", tempDir, "--config-file="+configFile)
-	var errb bytes.Buffer
-	mountCmd.Stderr = &errb
-	_, err := mountCmd.Output()
-	suite.NotEqual(nil, err)
-	fmt.Println(errb.String())
-	suite.NotEqual(0, len(errb.String()))
-	suite.Contains(errb.String(), "mount directory does not exists")
+	if runtime.GOOS == "windows" {
+		os.Mkdir(mntDir, 0777)
+		mountCmd := exec.Command(blobfuseBinary, "mount", mntDir, "--config-file="+configFile)
+		var errb bytes.Buffer
+		mountCmd.Stderr = &errb
+		_, err := mountCmd.Output()
+		suite.NotEqual(nil, err)
+		fmt.Println(errb.String())
+		suite.NotEqual(0, len(errb.String()))
+		suite.Contains(errb.String(), "Cannot create WinFsp-FUSE file system")
 
-	// list blobfuse mounted directories
-	cliOut := listBlobfuseMounts(suite)
-	suite.Equal(0, len(cliOut))
+		// list blobfuse mounted directories
+		cliOut := listBlobfuseMounts(suite)
+		suite.Equal(0, len(cliOut))
 
-	// unmount
-	blobfuseUnmount(suite, "Nothing to unmount")
+		// unmount
+		blobfuseUnmount(suite, "Nothing to unmount")
+		os.RemoveAll(mntDir)
+	} else {
+		tempDir := filepath.Join(mntDir, "tempdir")
+		mountCmd := exec.Command(blobfuseBinary, "mount", tempDir, "--config-file="+configFile)
+		var errb bytes.Buffer
+		mountCmd.Stderr = &errb
+		_, err := mountCmd.Output()
+		suite.NotEqual(nil, err)
+		fmt.Println(errb.String())
+		suite.NotEqual(0, len(errb.String()))
+		suite.Contains(errb.String(), "mount directory does not exists")
+
+		// list blobfuse mounted directories
+		cliOut := listBlobfuseMounts(suite)
+		suite.Equal(0, len(cliOut))
+
+		// unmount
+		blobfuseUnmount(suite, "Nothing to unmount")
+	}
 }
 
 // mount failure test where the mount directory is not empty
@@ -151,13 +180,23 @@ func (suite *mountSuite) TestMountDirNotEmpty() {
 	suite.NotEqual(nil, err)
 	fmt.Println(errb.String())
 	suite.NotEqual(0, len(errb.String()))
-	suite.Contains(errb.String(), "mount directory is not empty")
+
+	if runtime.GOOS == "windows" {
+		suite.Contains(errb.String(), "Cannot create WinFsp-FUSE file system")
+	} else {
+		suite.Contains(errb.String(), "mount directory is not empty")
+	}
 
 	// list blobfuse mounted directories
 	cliOut := listBlobfuseMounts(suite)
 	suite.Equal(0, len(cliOut))
 
 	os.RemoveAll(tempDir)
+
+	// Mount directory cannot exist on Windows so we need to remove it
+	if runtime.GOOS == "windows" {
+		os.RemoveAll(mntDir)
+	}
 
 	// unmount
 	blobfuseUnmount(suite, "Nothing to unmount")
@@ -232,6 +271,13 @@ func (suite *mountSuite) TestEnvVarMountFailure() {
 
 // mount test using environment variables for mounting
 func (suite *mountSuite) TestEnvVarMount() {
+	// TODO: Allow test to run on Windows.
+	if runtime.GOOS == "windows" {
+		suite.Equal(0, 0)
+		fmt.Println("Skipping tests on windows. We should enable this in the future.")
+		return
+	}
+
 	// read config file
 	configData, err := os.ReadFile(configFile)
 	suite.Equal(nil, err)
@@ -483,9 +529,14 @@ func TestMain(m *testing.M) {
 	if err != nil {
 		fmt.Println("Could not cleanup mount directory before testing")
 	}
-	os.Mkdir(mntDir, 0777)
+
+	if runtime.GOOS != "windows" {
+		os.Mkdir(mntDir, 0777)
+	}
 
 	m.Run()
 
-	os.RemoveAll(mntDir)
+	if runtime.GOOS != "windows" {
+		os.RemoveAll(mntDir)
+	}
 }
